+++ Src/CSharpier.Cli/ConsoleLogger.cs
@@ -110,9 +110,7 @@ internal class ConsoleLogger : ILogger
     }
 
     public IDisposable BeginScope<TState>(TState state)
-#if NET7_0_OR_GREATER
         where TState : notnull
-#endif
     {
         throw new NotImplementedException();
     }