+++ Src/CSharpier.Cli/Server/ServerFormatter.cs
@@ -5,24 +5,40 @@ using System.Net.NetworkInformation;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Hosting.Server.Features;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using NReco.Logging.File;
 
 internal static class ServerFormatter
 {
     public static async Task<int> StartServer(
         int? port,
         ConsoleLogger logger,
-        string? actualConfigPath,
-        CancellationToken cancellationToken
+        string? actualConfigPath
     )
     {
         var builder = WebApplication.CreateBuilder();
         builder.WebHost.ConfigureKestrel(
             (_, serverOptions) =>
             {
-                serverOptions.Listen(IPAddress.Loopback, 0);
+                serverOptions.Listen(IPAddress.Loopback, port ?? 0);
             }
         );
+        builder.Logging.ClearProviders();
+        var values = new Dictionary<string, string?>
+        {
+            ["Logging:File:MaxRollingFiles"] = "1",
+            ["Logging:File:FileSizeLimitBytes"] = "10000",
+        };
+        builder.Configuration.AddInMemoryCollection(values);
+        builder.Services.AddLogging(loggingBuilder =>
+        {
+            loggingBuilder.AddFile(
+                Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "server.log"),
+                append: true
+            );
+        });
 
         var app = builder.Build();
         app.Lifetime.ApplicationStarted.Register(() =>
@@ -37,7 +53,11 @@ internal static class ServerFormatter
                 logger.LogInformation("Started on " + uri.Port);
             }
         });
-        var service = new CSharpierServiceImplementation(actualConfigPath, logger);
+        var service = new CSharpierServiceImplementation(
+            actualConfigPath,
+            // we want any further logging to happen in the file log, not out to the console
+            app.Services.GetRequiredService<ILogger<CSharpierServiceImplementation>>()
+        );
         app.MapPost(
             "/format",
             (FormatFileParameter formatFileDto, CancellationToken cancellationToken) =>