+++ Src/CSharpier/SyntaxPrinter/SeparatedSyntaxList.cs
@@ -7,7 +7,8 @@ internal static class SeparatedSyntaxList
         Func<T, FormattingContext, Doc> printFunc,
         Doc afterSeparator,
         FormattingContext context,
-        int startingIndex = 0
+        int startingIndex = 0,
+        Doc? trailingSeparator = null
     )
         where T : SyntaxNode
     {
@@ -16,16 +17,25 @@ internal static class SeparatedSyntaxList
         {
             docs.Add(printFunc(list[x], context));
 
+            var isTrailingSeparator = x == list.Count - 1;
+
             if (x >= list.SeparatorCount)
             {
+                if (isTrailingSeparator && trailingSeparator != null)
+                {
+                    docs.Add(trailingSeparator);
+                }
+
                 continue;
             }
 
-            var isTrailingSeparator = x == list.Count - 1;
-
-            docs.Add(Token.Print(list.GetSeparator(x), context));
-            if (!isTrailingSeparator)
+            if (isTrailingSeparator)
             {
+                docs.Add(Doc.IfBreak(Token.Print(list.GetSeparator(x), context), Doc.Null));
+            }
+            else
+            {
+                docs.Add(Token.Print(list.GetSeparator(x), context));
                 docs.Add(afterSeparator);
             }
         }