+++ Src/CSharpier/SyntaxPrinter/SyntaxNodePrinters/ListPattern.cs
@@ -8,7 +8,13 @@ internal static class ListPattern
             Token.Print(node.OpenBracketToken, context),
             Doc.Indent(
                 Doc.SoftLine,
-                SeparatedSyntaxList.Print(node.Patterns, Node.Print, Doc.Line, context)
+                SeparatedSyntaxList.Print(
+                    node.Patterns,
+                    Node.Print,
+                    Doc.Line,
+                    context,
+                    trailingSeparator: TrailingComma.Print(node.CloseBracketToken, context)
+                )
             ),
             Doc.SoftLine,
             Token.Print(node.CloseBracketToken, context),