+++ Src/CSharpier.VisualStudio/ChangeLog.md
@@ -1,4 +1,8 @@
-﻿## [1.7.3]
+﻿## [1.7.4]
+- Support for semver
+- Only use CSharpier Server on 0.29.0+
+
+## [1.7.3]
 - If CSharpier doesn't respond when trying to find a port, then try to find a port in extension
 - Add option to bypass csharpier server.